feat: implementada policy de Fornecedor.
fix: correções de bugs e erros em relação a endpoint de proteção das policy.

+++ app/Domains/Inventario/Controllers/FornecedorController.php
@@ -8,6 +8,7 @@ use App\Domains\Inventario\Requests\Fornecedor\UpdateFornecedorRequest;
 use App\Domains\Inventario\Resources\FornecedorResource;
 use App\Domains\Inventario\Services\FornecedorService;
 use App\Http\Controllers\Controller;
+use App\Models\Fornecedor;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
@@ -21,28 +22,33 @@ class FornecedorController extends Controller
 
     public function index(): AnonymousResourceCollection
     {
-        return FornecedorResource::collection($this->fornecedorService->findAll());
+        $this->authorize('viewAny', Fornecedor::class);
+        return FornecedorResource::collection($this->fornecedorService->findAll(auth()->user()));
     }
 
     public function show(int $id): FornecedorResource
     {
+        $this->authorize('view', Fornecedor::class);
         $fornecedor = $this->fornecedorService->find($id);
         return new FornecedorResource($fornecedor);
     }
 
     public function store(StoreFornecedorRequest $request): FornecedorResource
     {
+        $this->authorize('createForRestaurante', Fornecedor::class);
         $fornecedor = $this->fornecedorService->create($request->validated());
         return new FornecedorResource($fornecedor);
     }
 
     public function update(UpdateFornecedorRequest $request, int $id): FornecedorResource
     {
+        $this->authorize('update', Fornecedor::class);
         $fornecedor = $this->fornecedorService->update($id, $request->validated());
         return new FornecedorResource($fornecedor);
     }
     public function destroy(int $id): JsonResponse
     {
+        $this->authorize('delete', Fornecedor::class);
         $this->fornecedorService->delete($id);
         return response()->json(["message" => "Fornecedor deletado com sucesso!"], 200);
     }
